added approximation of poisson with normal for large lambdas
When lambda (read count) is larger than 1000, a normal distribution will
be used from now on for integration. This improves the robustness of the
exppression parts of model. It was previously fixed ad-hoc by assigning
all the probability mass to the right-most bin, which appears dubious to
me. The expression part appears to be ok otherwise. The 2way model still
shows low correlation with the SotA methods combined with Fisher's.